+++ client/capi/gain_holo/c_api.cpp
@@ -3,7 +3,7 @@
 // Created Date: 02/07/2018
 // Author: Shun Suzuki
 // -----
-// Last Modified: 01/05/2021
+// Last Modified: 06/05/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2018-2020 Hapis Lab. All rights reserved.
@@ -13,17 +13,77 @@
 #include "./gain_holo.h"
 #include "gain/holo.hpp"
 
-void AUTDHoloGain(void** gain, float* points, float* amps, const int32_t size, int32_t method, void* params) {
+namespace {
+std::vector<autd::Vector3> pack_foci(const float* const points, const int32_t size) {
   std::vector<autd::Vector3> holo;
-  std::vector<autd::Float> amps_;
+  holo.reserve(size);
   for (auto i = 0; i < size; i++) {
     const auto x = static_cast<autd::Float>(points[3 * i]);
     const auto y = static_cast<autd::Float>(points[3 * i + 1]);
     const auto z = static_cast<autd::Float>(points[3 * i + 2]);
     holo.emplace_back(autd::Vector3(x, y, z));
+  }
+  return holo;
+}
+
+std::vector<autd::Float> pack_amps(const float* const amps, const int32_t size) {
+  std::vector<autd::Float> amps_;
+  amps_.reserve(size);
+  for (auto i = 0; i < size; i++) {
     amps_.emplace_back(static_cast<autd::Float>(amps[i]));
   }
-  const auto method_ = static_cast<autd::gain::holo::OPT_METHOD>(method);
-  auto* g = GainCreate(autd::gain::holo::HoloGain<autd::gain::holo::Eigen3Backend>::Create(holo, amps_, method_, params));
+  return amps_;
+}
+}  // namespace
+
+void AUTDHoloGainSDP(void** gain, float* points, float* amps, const int32_t size, const float alpha, const float lambda, const uint64_t repeat,
+                     const bool normalize) {
+  const auto holo = pack_foci(points, size);
+  const auto amps_ = pack_amps(amps, size);
+  auto* g = GainCreate(autd::gain::holo::HoloGainSDP<autd::gain::holo::Eigen3Backend>::Create(holo, amps_, static_cast<autd::Float>(alpha),
+                                                                                              static_cast<autd::Float>(lambda), repeat, normalize));
+  *gain = g;
+}
+
+void AUTDHoloGainEVD(void** gain, float* points, float* amps, const int32_t size, const float gamma, const bool normalize) {
+  const auto holo = pack_foci(points, size);
+  const auto amps_ = pack_amps(amps, size);
+  auto* g =
+      GainCreate(autd::gain::holo::HoloGainEVD<autd::gain::holo::Eigen3Backend>::Create(holo, amps_, static_cast<autd::Float>(gamma), normalize));
+  *gain = g;
+}
+
+void AUTDHoloGainNaive(void** gain, float* points, float* amps, const int32_t size) {
+  const auto holo = pack_foci(points, size);
+  const auto amps_ = pack_amps(amps, size);
+  auto* g = GainCreate(autd::gain::holo::HoloGainNaive<autd::gain::holo::Eigen3Backend>::Create(holo, amps_));
+  *gain = g;
+}
+
+void AUTDHoloGainGS(void** gain, float* points, float* amps, const int32_t size, const uint64_t repeat) {
+  const auto holo = pack_foci(points, size);
+  const auto amps_ = pack_amps(amps, size);
+  auto* g = GainCreate(autd::gain::holo::HoloGainGS<autd::gain::holo::Eigen3Backend>::Create(holo, amps_, repeat));
+  *gain = g;
+}
+
+void AUTDHoloGainGSPAT(void** gain, float* points, float* amps, const int32_t size, const uint64_t repeat) {
+  const auto holo = pack_foci(points, size);
+  const auto amps_ = pack_amps(amps, size);
+  auto* g = GainCreate(autd::gain::holo::HoloGainGSPAT<autd::gain::holo::Eigen3Backend>::Create(holo, amps_, repeat));
+  *gain = g;
+}
+
+void AUTDHoloGainLM(void** gain, float* points, float* amps, const int32_t size, const float eps_1, const float eps_2, const float tau,
+                    const uint64_t k_max, float* initial, const int32_t initial_size) {
+  const auto holo = pack_foci(points, size);
+  const auto amps_ = pack_amps(amps, size);
+
+  std::vector<autd::Float> initial_;
+  initial_.reserve(initial_size);
+  for (auto i = 0; i < initial_size; i++) initial_.emplace_back(static_cast<autd::Float>(initial[i]));
+
+  auto* g = GainCreate(autd::gain::holo::HoloGainLM<autd::gain::holo::Eigen3Backend>::Create(
+      holo, amps_, static_cast<autd::Float>(eps_1), static_cast<autd::Float>(eps_2), static_cast<autd::Float>(tau), k_max, initial_));
   *gain = g;
 }