+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.9.3
+* Add `Rebuild()` method to Gain
+* Change the declaration of HoloGain. Selection by OPT_METHOD is deleted and introduced HoloGainSDP, HoloGainEVD, HoloGainNaive, HoloGainGS, HoloGainGSPAT, HoloGainLM classes.
+
 # 0.9.2
 * Make double-precision float default
     * `USE_DOUBLE_AUTD` was removed, and `USE_SINGLE_FLOAT_AUTD` was added